feat: creating discovery page layout for f14
